requirements.txt
appdirs==1.4.4
backcall==0.2.0
black==21.5b1
click==8.0.0
cycler==0.10.0
decorator==5.0.7
fancycompleter==0.9.1
ipython==7.23.1
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
kiwisolver==1.3.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mypy-extensions==0.4.3
numpy==1.20.3
pandas==1.2.4
parso==0.8.2
pathspec==0.8.1
pdbpp==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
PyMeasure==0.9.0
pyparsing==2.4.7
pyqtgraph==0.12.1
pyrepl==0.9.0
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
PyVISA==1.11.3
regex==2021.4.4
six==1.16.0
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.10.0.0
wcwidth==0.2.5
wmctrl==0.4
